backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-dotenv==1.0.0
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
google-generativeai>=0.8.0
chromadb==0.5.0
sentence-transformers==3.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
aiofiles==23.2.1
aiosqlite==0.20.0
